adapt changes to directory structure on main branch for merge
