--- src/backend/data_acquisition/config.json
@@ -1,32 +1,0 @@
-{
-    "archives_targets": [
-        {
-            "keywords": [
-                "nutrition",
-                "health",
-                "food as medicine"
-            ],
-            "urls": [
-                "http://arxiv.org",
-                "https://pubmed.ncbi.nlm.nih.gov/"
-            ]
-        }
-    ],
-    "podcasts_targets": [
-        {
-            "keywords": [
-                "podcast",
-                "health"
-            ],
-            "urls": [
-                "https://peterattiamd.com/podcast/archive/"
-            ]
-        }
-    ],
-    "youtube_channel_targets": [
-        {
-            "name": "Dr William Li",
-            "url": "https://www.youtube.com/@DrWilliamLi"
-        }
-    ]
-}
