--- src/backend/data_acquisition/scrapers/nutritionfacts/main.py
@@ -1,5 +1,0 @@
-from nutritionfacts import scrape_blogs
-
-if __name__ == "__main__":
-    base_url = 'https://nutritionfacts.org/blog/'
-    scrape_blogs(base_url)
