--- src/backend/data_acquisition/scrapers/podcast/vosk-model-small-en-us-0.15/README
@@ -1,9 +1,0 @@
-US English model for mobile Vosk applications
-
-Copyright 2020 Alpha Cephei Inc
-
-Accuracy: 10.38 (tedlium test) 9.85 (librispeech test-clean)
-Speed: 0.11xRT (desktop)
-Latency: 0.15s (right context)
-
-
