--- src/backend/data_acquisition/scrapers/podcast/vosk-model-small-en-us-0.15/conf/mfcc.conf
@@ -1,7 +1,0 @@
---sample-frequency=16000
---use-energy=false
---num-mel-bins=40
---num-ceps=40
---low-freq=20
---high-freq=7600
---allow-downsample=true
