--- src/backend/data_acquisition/scrapers/podcast/vosk-model-small-en-us-0.15/conf/model.conf
@@ -1,10 +1,0 @@
---min-active=200
---max-active=3000
---beam=10.0
---lattice-beam=2.0
---acoustic-scale=1.0
---frame-subsampling-factor=3
---endpoint.silence-phones=1:2:3:4:5:6:7:8:9:10
---endpoint.rule2.min-trailing-silence=0.5
---endpoint.rule3.min-trailing-silence=0.75
---endpoint.rule4.min-trailing-silence=1.0
