--- src/backend/data_acquisition/scrapers/podcast/vosk-model-small-en-us-0.15/graph/phones/word_boundary.int
@@ -1,166 +1,0 @@
-1 nonword
-2 begin
-3 end
-4 internal
-5 singleton
-6 nonword
-7 begin
-8 end
-9 internal
-10 singleton
-11 begin
-12 end
-13 internal
-14 singleton
-15 begin
-16 end
-17 internal
-18 singleton
-19 begin
-20 end
-21 internal
-22 singleton
-23 begin
-24 end
-25 internal
-26 singleton
-27 begin
-28 end
-29 internal
-30 singleton
-31 begin
-32 end
-33 internal
-34 singleton
-35 begin
-36 end
-37 internal
-38 singleton
-39 begin
-40 end
-41 internal
-42 singleton
-43 begin
-44 end
-45 internal
-46 singleton
-47 begin
-48 end
-49 internal
-50 singleton
-51 begin
-52 end
-53 internal
-54 singleton
-55 begin
-56 end
-57 internal
-58 singleton
-59 begin
-60 end
-61 internal
-62 singleton
-63 begin
-64 end
-65 internal
-66 singleton
-67 begin
-68 end
-69 internal
-70 singleton
-71 begin
-72 end
-73 internal
-74 singleton
-75 begin
-76 end
-77 internal
-78 singleton
-79 begin
-80 end
-81 internal
-82 singleton
-83 begin
-84 end
-85 internal
-86 singleton
-87 begin
-88 end
-89 internal
-90 singleton
-91 begin
-92 end
-93 internal
-94 singleton
-95 begin
-96 end
-97 internal
-98 singleton
-99 begin
-100 end
-101 internal
-102 singleton
-103 begin
-104 end
-105 internal
-106 singleton
-107 begin
-108 end
-109 internal
-110 singleton
-111 begin
-112 end
-113 internal
-114 singleton
-115 begin
-116 end
-117 internal
-118 singleton
-119 begin
-120 end
-121 internal
-122 singleton
-123 begin
-124 end
-125 internal
-126 singleton
-127 begin
-128 end
-129 internal
-130 singleton
-131 begin
-132 end
-133 internal
-134 singleton
-135 begin
-136 end
-137 internal
-138 singleton
-139 begin
-140 end
-141 internal
-142 singleton
-143 begin
-144 end
-145 internal
-146 singleton
-147 begin
-148 end
-149 internal
-150 singleton
-151 begin
-152 end
-153 internal
-154 singleton
-155 begin
-156 end
-157 internal
-158 singleton
-159 begin
-160 end
-161 internal
-162 singleton
-163 begin
-164 end
-165 internal
-166 singleton
