--- src/backend/data_acquisition/scrapers/podcast/vosk-model-small-en-us-0.15/ivector/online_cmvn.conf
@@ -1,1 +1,0 @@
-# configuration file for apply-cmvn-online, used in the script ../local/run_online_decoding.sh
