--- src/backend/data_acquisition/scrapers/podcast/vosk-model-small-en-us-0.15/ivector/splice.conf
@@ -1,2 +1,0 @@
---left-context=3
---right-context=3
