--- src/backend/data_acquisition/scrapers/youtube/index.py
@@ -1,4 +1,0 @@
-from scrapper.you_tube import YouTube
-
-if __name__ == '__main__':
-  print(YouTube.get_video_transcript('uF5NAj83wc0', 'en'))
